change `alwaysThrows3` to int (originally float)

+++ course-02242-examples/src/executables/java/eu/bogoe/dtu/exceptional/Arithmetics.java
@@ -17,7 +17,7 @@ public class Arithmetics {
 	}
 
 	@Case
-	public static float alwaysThrows3(float i, float j) {
+	public static int alwaysThrows3(int i, int j) {
 		return i / j;
 	}
